+++ core/services.go
@@ -971,8 +971,20 @@ func (app *Application) deletePost(clientID string, userID string, groupID strin
 
 // TODO this logic needs to be refactored because it's over complicated!
 func (app *Application) synchronizeAuthman(clientID string, stemNames []string) error {
+
+	if app.authmanSyncInProgress {
+		log.Printf("Another Authman sync process is running")
+		return fmt.Errorf("another Authman sync process is running")
+	}
+
 	log.Printf("Global Authman synchronization started")
-	defer log.Printf("Global Authman synchronization finished")
+
+	app.authmanSyncInProgress = true
+	finishAuthmanSync := func() {
+		app.authmanSyncInProgress = false
+		log.Printf("Global Authman synchronization finished")
+	}
+	defer finishAuthmanSync()
 
 	if len(stemNames) > 0 {
 		for _, stemName := range stemNames {
@@ -991,9 +1003,28 @@ func (app *Application) synchronizeAuthman(clientID string, stemNames []string) 
 					if storedGiesGroup == nil {
 						title, adminUINs := giesGroup.GetGroupPettyTitleAndAdmins()
 
-						var members []model.Member
+						defaultAdminsMapping := map[string]bool{}
 						if len(adminUINs) > 0 {
-							members = app.buildMembersByExternalIDs(clientID, adminUINs, "admin")
+							for _, adminUIN := range adminUINs {
+								defaultAdminsMapping[adminUIN] = true
+							}
+						}
+						if len(app.config.AuthmanAdminUINList) > 0 {
+							for _, externalID := range app.config.AuthmanAdminUINList {
+								defaultAdminsMapping[externalID] = true
+							}
+						}
+
+						var constructedAdminUINs []string
+						if len(defaultAdminsMapping) > 0 {
+							for key := range defaultAdminsMapping {
+								constructedAdminUINs = append(constructedAdminUINs, key)
+							}
+						}
+
+						var members []model.Member
+						if len(constructedAdminUINs) > 0 {
+							members = app.buildMembersByExternalIDs(clientID, constructedAdminUINs, "admin")
 						}
 
 						emptyText := ""
@@ -1028,9 +1059,8 @@ func (app *Application) synchronizeAuthman(clientID string, stemNames []string) 
 										storedGiesGroup.Members[index].DateUpdated = &now
 										groupUpdated = true
 										break
-									} else {
-										found = true
 									}
+									found = true
 								}
 							}
 							if !found {
@@ -1132,12 +1162,6 @@ func (app *Application) synchronizeAuthmanGroup(clientID string, authmanGroup *m
 	defer log.Printf("Authman synchronization for group %s finished", authmanGroup.Title)
 
 	defaultAdminsMapping := map[string]bool{}
-	if len(app.config.AuthmanAdminUINList) > 0 {
-		for _, externalID := range app.config.AuthmanAdminUINList {
-			defaultAdminsMapping[externalID] = true
-		}
-	}
-
 	admins := authmanGroup.GetAllAdminMembers()
 	if len(admins) > 0 {
 		for _, admin := range admins {
@@ -1275,7 +1299,7 @@ func (app *Application) synchronizeAuthmanGroup(clientID string, authmanGroup *m
 					}
 					if !found {
 						members = append(members, member)
-						log.Printf("add remaining admin user(%s, %s, %s) to '%s'", member.UserID, member.Name, member.Email, authmanGroup.Title)
+						log.Printf("add remaining admin user(%s, %s, %s, %s) to '%s'", member.UserID, member.ExternalID, member.Name, member.Email, authmanGroup.Title)
 					}
 				} else if _, ok := newExternalMembersMapping[member.ExternalID]; !ok {
 					log.Printf("User(%s, %s) will be removed as a member of '%s', because it's not defined in Authman group", member.ExternalID, member.Name, authmanGroup.Title)